Use `org.bukkit.plugin.Plugin` in constructors
Due to the way PlaceholderAPI loads expansions, it doesn't seem to be
possible to make the main expansion class tackle the Plugin down casts,
so this commit moves them into the various resolver classes. This is
probably not the most elegant way to tackle the problem, because it does
push an annoying "keep track of the parameter order" problem on the main
expansion class, but it solves the problem of MobArenaStats being a hard
dependency on the expansion, which it shouldn't be.