Add `null` tail safeguards.
ArenaResolver and StatsResolver both throw NullPointerExceptions if the
expansion is asked to resolve the "bare" placeholders for each resolver,
i.e. `mobarena_arena` or `mobarena_stats`. This is obviously incorrect
behavior, as all the bare placeholders are invalid.

This commit adds an early return on `null` tails at the beginning of the
`resolve()` method of the two resolvers. It also relocates the existing
check in PlayerResolver for consistency.