Juggle linebreaks in `pom.xml`.
We basically want each section to be separate from the next one, but the
first and last sections "stick" to their parent section. The exception
is the `project` tag, whose outer shell we do want some spacing from.
This sort of resembles the code style in the Java classes, which should
make it somewhat familiar.