Add Vault as a test dependency.
Because MobArena uses Vault, and its interface exposes a component from
the Vault project, Mockito requires that we can resolve the Vault stuff
to go through with the mocking of the MobArena plugin class. Yeah...

To get rid of this dependency in the future, we need MobArena to ditch
its `getEconomy()` method. That is unlikely to happen immediately, so we
have to make do with this setup. Hopefully the main MobArena interface
will look a bit different after the Sessions Rework :)