Add a skeleton test suite for each resolver.
Introduces a test class for each of the three resolvers so we can start
using a test-driven approach for further development. This should help
make the code a bit more robust and the design a bit more deliberate in
terms of "what to do given X".

The following commits should complete the picture.