Fix PlayerResolver behavior.
This commit fixes the incorrect behavior of PlayerResolver in terms of
placeholders that are _invalid_ vs. _empty_. The implementation is very
"minimal effort" for illustrative purposes only. It aims to show _just_
the necessary changes from the current behavior to fit the specification
that PlaceholderAPI Expansions are expected to follow. This means that a
few oddities are left in, and the accompanying test suite also contains
a few oddities to match (such as the `Player` instance returning itself
when its `getPlayer()` method is called). A cleanup commit follows.

An _invalid_ placeholder is a placeholder that doesn't "exist", e.g. the
`mobarena_player_cheese` placeholder alluded to in the test suite. Such
a placeholder must return null to trigger errors downstream, rather than
a non-null string, which _should_ be reserved for valid placeholders.

An _empty_ placeholder is a valid placeholder (it _does_ exist), but one
that cannot possibly produce a meaningful value in the current context.
An example would be `mobarena_player_class`, which only makes sense when
the given player is actually in an arena session.