Refactor PlayerResolver.
This commit removes the extra `switch` statement introduced previously.

This new solution revolves around higher-order functions that allow us
to defer the value extraction until after the player- and arena-centric
validation logic has been evaluated.

A couple of helper methods are introduced to help resolve the player and
arena instances. This should help with extensibility if we want to add
more placeholders later down the road.

Note that due to the semi-solid test suite implemented previously, there
are no changes made elsewhere in the code base - a clean refactoring :)