Fix ArenaResolver behavior.
This commit fixes the incorrect behavior of ArenaResolver in terms of
placeholders that are _invalid_ vs. _empty_ similar to previous work in
commit fa65c51 for PlayerResolver.

ArenaResolver has an extra variable in the form of a _key_, which can be
either an arena slug or the sentinel value `$current`, which means the
arena that the _target_ player is currently participating in. This gives
us an extra thing to think about, but overall, the work here is pretty
much the same as in the previous commit.

In the following commit(s), we'll clean this bad boy up!