Fix StatsResolver behavior.
Similar to the effort with the other two resolvers, this commit fixes
the StatsResolver behavior to be consistent in terms of _empty_ and
_invalid_ placeholders.

This resolver is a little different in that it "houses" three different
responsibilities lumped together. The heterogeneity of the test suite
shows just how separate the concerns are, perhaps a good indicator of a
"refactorable" component...

At any rate, as with the other two efforts, the goal here is purely that
of behavior fixing, and a cleanup of the duplicate `switch` statements
is right around the corner from here.