Extract arena resolution from resolvers.
Moves the "look up arena by player/name" into a separate component that
deals _just_ with that responsibility. The lookup logic is "noisy" in
the context of the resolvers, and it doesn't have to be.

Note how the resolvers (and their test suites) get simpler and a little
less crowded. This is a good thing!