Fix name of arena slug placeholder.
This is really just a matter of semantics, since we'll add the name as
well. It's probably not a bad idea to bring whatever amount of focus we
can to the difference, since the slugs are so heavily used by MobArena
internally, and this kind of stuff walks a line between representational
logic and actual integration logic. Integrations will want to use slugs.